Struct oorandom::Rand64

pub struct Rand64 { /* fields omitted */ }

A PRNG producing a 64-bit output.

The current implementation is PCG-XSH-RR.

Implementations

impl Rand64

pub const DEFAULT_INC: u128

The default value for increment.

The value used here is from the PCG default C implementation: http://www.pcg-random.org/download.html

pub fn new(seed: u128) -> Self

Creates a new PRNG with the given seed and a default increment.

pub fn new_inc(seed: u128, increment: u128) -> Self

Same as Rand32::new_inc()

pub fn state(&self) -> (u128, u128)

Returns the internal state of the PRNG. This allows you to save a PRNG and create a new one that will resume from the same spot in the sequence.

pub fn from_state(state: (u128, u128)) -> Self

Creates a new PRNG from a saved state from Rand32::state(). This is NOT quite the same as new_inc() because new_inc() does a little extra setup work to initialize the state.

pub fn rand_u64(&mut self) -> u64

Produces a random u64 in the range[0, u64::MAX].

pub fn rand_i64(&mut self) -> i64

Produces a random i64 in the range [i64::MIN, i64::MAX].

pub fn rand_float(&mut self) -> f64

Produces a random f64 in the range [0.0, 1.0).

pub fn rand_range(&mut self, range: Range<u64>) -> u64

Produces a random within the given bounds. Like any Range, it includes the lower bound and excludes the upper one.

This should be faster than Self::rand() % end + start, but the real advantage is it’s more convenient. Requires that range.end <= range.start.